+++ negotiator.go
@@ -124,8 +124,6 @@ func negotiator(cfg StreamConfig) Negotiator {
 			}
 		}
 
-		// TODO: Check if the first token is a stream error (if so, unmarshal and
-		// return, otherwise pass the token into negotiateFeatures).
 		mask, rw, err = negotiateFeatures(ctx, s, data == nil, cfg.Features)
 		nState.doRestart = rw != nil
 		return mask, rw, nState, err